🌱 fix: split OperatorDrillDown and DriftDrillDown into hooks and parts
- OperatorDrillDown.tsx: 649 → 175 lines
- DriftDrillDown.tsx: 619 → 191 lines
- Added useOperatorDrillDown.ts and useDriftDrillDown.ts for data-fetch logic
- Added OperatorDrillDownHeader, OperatorTabBar to OperatorDrillDown.parts.tsx
- Added DriftDrillDownHeader, DriftTabBar to DriftDrillDown.parts.tsx
- Pure extraction; no behaviour change
