+++ web/src/components/drilldown/views/DriftDrillDown.tsx
@@ -1,16 +1,7 @@
-import { useState, useEffect, useRef } from 'react'
-import { useLocalAgent } from '../../../hooks/useLocalAgent'
-import { useDrillDownWebSocket } from '../../../hooks/useDrillDownWebSocket'
+import { useState } from 'react'
+import { Info, Diff, FileText, Stethoscope } from 'lucide-react'
 import { useDrillDownActions, useDrillDown } from '../../../hooks/useDrillDown'
 import { useMissions } from '../../../hooks/useMissions'
-import { ClusterBadge } from '../../ui/ClusterBadge'
-import {
-  GitBranch, Info, Loader2, Server, Stethoscope,
-  AlertTriangle, CheckCircle,
-  FileText, Layers, ArrowRight, Diff
-} from 'lucide-react'
-import { cn } from '../../../lib/cn'
-import { ConsoleAIIcon } from '../../ui/ConsoleAIIcon'
 import {
   AIActionBar,
   useModalAI,
@@ -24,6 +15,15 @@ import {
   getDriftSeverityStyle,
   getChangeTypeStyle,
 } from './drift-drilldown'
+import {
+  DriftDrillDownHeader,
+  DriftTabBar,
+  DriftOverviewTab,
+  DriftChangesTab,
+  DriftDiffTab,
+  DriftAITab,
+} from './DriftDrillDown.parts'
+import { useDriftDrillDown } from './useDriftDrillDown'
 
 export function DriftDrillDown({ data }: Props) {
   const { t } = useTranslation()
@@ -31,7 +31,6 @@ export function DriftDrillDown({ data }: Props) {
   const namespace = data.namespace as string | undefined
   const resourceName = data.resource as string | undefined
 
-  // Drift detection data
   const driftStatus = (data.status as string) || 'Unknown'
   const driftSeverity = (data.severity as string) || driftStatus
   const gitRepo = data.gitRepo as string | undefined
@@ -40,20 +39,16 @@ export function DriftDrillDown({ data }: Props) {
   const lastChecked = data.lastChecked as string | undefined
   const driftedResources = (data.driftedResources as number) || 0
 
-  const { isConnected: agentConnected } = useLocalAgent()
   const { drillToNamespace, drillToCluster, drillToPod, drillToDeployment } = useDrillDownActions()
   const { close: closeDrillDown } = useDrillDown()
   const { startMission } = useMissions()
 
   const [activeTab, setActiveTab] = useState<TabType>('overview')
-  const [changes, setChanges] = useState<DriftChange[] | null>(null)
-  const [changesLoading, setChangesLoading] = useState(false)
-  const [changesError, setChangesError] = useState<string | null>(null)
-  const [selectedChange, setSelectedChange] = useState<DriftChange | null>(null)
-  const [aiAnalysis] = useState<string | null>(null)
-  const [aiAnalysisLoading] = useState(false)
 
-  // Resource context for AI actions
+  const { changes, changesLoading, changesError, selectedChange, setSelectedChange } = useDriftDrillDown(
+    cluster, namespace,
+  )
+
   const resourceContext: ResourceContext = {
     kind: 'Custom',
     name: resourceName || 'GitOps Drift',
@@ -62,129 +57,17 @@ export function DriftDrillDown({ data }: Props) {
     status: driftStatus,
   }
 
-  // Check for issues
   const hasIssues = driftedResources > 0 || driftSeverity.toLowerCase() === 'high'
   const issues = hasIssues
     ? [{ name: 'Drift', message: `${driftedResources} drifted resources`, severity: 'warning' }]
     : []
 
-  // Use modal AI hook
   const { defaultAIActions, handleAIAction, isAgentConnected } = useModalAI({
     resource: resourceContext,
     issues,
-    additionalContext: {
-      gitRepo,
-      gitBranch,
-      gitPath,
-      driftedResources,
-    },
+    additionalContext: { gitRepo, gitBranch, gitPath, driftedResources },
   })
-  const { runKubectl } = useDrillDownWebSocket(cluster)
 
-
-  // Fetch drift details
-  const fetchDriftDetails = async () => {
-    if (!agentConnected || changes) return
-    setChangesLoading(true)
-    setChangesError(null)
-    try {
-      // Try to get drift from Flux or ArgoCD
-      // First try Flux Kustomization
-      if (namespace) {
-        const output = await runKubectl([
-          'get', 'kustomization', '-n', namespace, '-o', 'json'
-        ])
-        if (output) {
-          let ksList
-          try {
-            ksList = JSON.parse(output)
-          } catch {
-            setChanges([])
-            setChangesError('Failed to parse drift data')
-            return
-          }
-          const items = ksList.items || []
-          const driftChanges: DriftChange[] = []
-
-          for (const ks of items) {
-            // Check for drift annotations
-            if (ks.metadata?.annotations?.['kustomize.toolkit.fluxcd.io/driftDetection'] === 'enabled') {
-              const lastApplied = ks.status?.lastAppliedRevision
-              const lastHandled = ks.status?.lastHandledReconcileAt
-              if (lastApplied !== lastHandled) {
-                driftChanges.push({
-                  kind: 'Kustomization',
-                  name: ks.metadata?.name || 'Unknown',
-                  namespace: ks.metadata?.namespace,
-                  changeType: 'modified',
-                })
-              }
-            }
-          }
-
-          if (driftChanges.length > 0) {
-            setChanges(driftChanges)
-            return
-          }
-        }
-      }
-
-      // If no Flux data, try ArgoCD Applications
-      const argoOutput = await runKubectl([
-        'get', 'applications.argoproj.io', '-A', '-o', 'json'
-      ])
-      if (argoOutput) {
-        let appList
-        try {
-          appList = JSON.parse(argoOutput)
-        } catch {
-          setChanges([])
-          setChangesError('Failed to parse ArgoCD data')
-          return
-        }
-        const apps = appList.items || []
-        const driftChanges: DriftChange[] = []
-
-        for (const app of apps) {
-          const syncStatus = app.status?.sync?.status
-          const resources = app.status?.resources || []
-
-          if (syncStatus === 'OutOfSync') {
-            for (const res of resources) {
-              if (res.status === 'OutOfSync') {
-                driftChanges.push({
-                  kind: res.kind || 'Unknown',
-                  name: res.name || 'Unknown',
-                  namespace: res.namespace,
-                  changeType: 'modified',
-                })
-              }
-            }
-          }
-        }
-
-        setChanges(driftChanges)
-      } else {
-        setChanges([])
-      }
-    } catch (err) {
-      const errMsg = err instanceof Error ? err.message : 'Failed to fetch drift details'
-      setChanges([])
-      setChangesError(errMsg)
-    }
-    setChangesLoading(false)
-  }
-
-  // Track if we've already loaded data
-  const hasLoadedRef = useRef(false)
-
-  useEffect(() => {
-    if (!agentConnected || hasLoadedRef.current) return
-    hasLoadedRef.current = true
-    fetchDriftDetails()
-  }, [agentConnected, fetchDriftDetails])
-
-  // Navigate to resource
   const handleResourceClick = (change: DriftChange) => {
     if (change.kind === 'Pod' && change.namespace) {
       drillToPod(cluster, change.namespace, change.name)
@@ -193,9 +76,9 @@ export function DriftDrillDown({ data }: Props) {
     } else if (change.namespace) {
       drillToNamespace(cluster, change.namespace)
     }
+    setSelectedChange(change)
   }
 
-  // Start AI diagnosis
   const handleDiagnose = () => {
     const prompt = `Analyze GitOps drift for cluster "${cluster}".
 
@@ -226,26 +109,18 @@ Please:
    - "Should I check for drift in other namespaces?"
    - "All done"`
 
-    closeDrillDown() // Close panel so mission sidebar is visible
+    closeDrillDown()
     startMission({
       title: `Analyze GitOps Drift: ${cluster}`,
       description: `Investigate configuration drift between Git and cluster`,
       type: 'troubleshoot',
       cluster,
       initialPrompt: prompt,
-      context: {
-        kind: 'Drift',
-        name: 'GitOps Drift Analysis',
-        namespace,
-        cluster,
-        gitRepo,
-        driftedResources,
-      },
+      context: { kind: 'Drift', name: 'GitOps Drift Analysis', namespace, cluster, gitRepo, driftedResources },
     })
   }
 
   const severityStyle = getDriftSeverityStyle(driftSeverity)
-  const SeverityIcon = severityStyle.icon
 
   const TABS: { id: TabType; label: string; icon: typeof Info }[] = [
     { id: 'overview', label: t('drilldown.tabs.overview'), icon: Info },
@@ -256,45 +131,15 @@ Please:
 
   return (
     <div className="flex flex-col h-full -m-6">
-      {/* Header */}
-      <div className="px-6 pt-6 pb-4">
-        <div className="flex items-center justify-between">
-          <div className="flex items-center gap-6 text-sm">
-            {namespace && (
-              <button
-                onClick={() => drillToNamespace(cluster, namespace)}
-                className="flex items-center gap-2 hover:bg-purple-500/10 border border-transparent hover:border-purple-500/30 px-3 py-1.5 rounded-lg transition-all group cursor-pointer"
-              >
-                <Layers className="w-4 h-4 text-purple-400" />
-                <span className="text-muted-foreground">{t('drilldown.fields.namespace')}</span>
-                <span className="font-mono text-purple-400 group-hover:text-purple-300 transition-colors">{namespace}</span>
-                <svg className="w-3 h-3 text-purple-400/70 group-hover:text-purple-400 transition-colors" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-                </svg>
-              </button>
-            )}
-            <button
-              onClick={() => drillToCluster(cluster)}
-              className="flex items-center gap-2 hover:bg-blue-500/10 border border-transparent hover:border-blue-500/30 px-3 py-1.5 rounded-lg transition-all group cursor-pointer"
-            >
-              <Server className="w-4 h-4 text-blue-400" />
-              <span className="text-muted-foreground">{t('drilldown.fields.cluster')}</span>
-              <ClusterBadge cluster={cluster.split('/').pop() || cluster} size="sm" />
-              <svg className="w-3 h-3 text-blue-400/70 group-hover:text-blue-400 transition-colors" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-              </svg>
-            </button>
-          </div>
+      <DriftDrillDownHeader
+        cluster={cluster}
+        namespace={namespace}
+        driftSeverity={driftSeverity}
+        severityStyle={severityStyle}
+        onNamespaceClick={namespace ? () => drillToNamespace(cluster, namespace) : undefined}
+        onClusterClick={() => drillToCluster(cluster)}
+      />
 
-          {/* Status badge */}
-          <span className={cn('px-2.5 py-1 rounded-lg text-xs font-medium flex items-center gap-1', severityStyle.bg, severityStyle.text, 'border', severityStyle.border)}>
-            <SeverityIcon className="w-3 h-3" />
-            {driftSeverity === 'None' || driftStatus === 'Synced' ? 'In Sync' : 'Drifted'}
-          </span>
-        </div>
-      </div>
-
-      {/* AI Action Bar */}
       <div className="px-6 pb-4">
         <AIActionBar
           resource={resourceContext}
@@ -305,270 +150,40 @@ Please:
         />
       </div>
 
-      {/* Tabs */}
-      <div className="border-b border-border px-6">
-        <div className="flex gap-1">
-          {TABS.map((tab) => {
-            const Icon = tab.icon
-            return (
-              <button
-                key={tab.id}
-                onClick={() => setActiveTab(tab.id)}
-                className={cn(
-                  'px-4 py-2 text-sm font-medium flex items-center gap-2 border-b-2 transition-colors',
-                  activeTab === tab.id
-                    ? 'text-primary border-primary'
-                    : 'text-muted-foreground border-transparent hover:text-foreground hover:border-border'
-                )}
-              >
-                <Icon className="w-4 h-4" />
-                {tab.label}
-              </button>
-            )
-          })}
-        </div>
-      </div>
+      <DriftTabBar tabs={TABS} activeTab={activeTab} onSelect={(id) => setActiveTab(id as TabType)} />
 
-      {/* Tab Content */}
       <div className="flex-1 overflow-y-auto p-6 space-y-6">
         {activeTab === 'overview' && (
-          <div className="space-y-6">
-            {/* Drift Status Card */}
-            <div className={cn(
-              'p-4 rounded-lg border',
-              driftedResources > 0 ? 'bg-linear-to-r from-red-500/10 to-orange-500/10 border-red-500/20' : 'bg-linear-to-r from-green-500/10 to-green-500/10 border-green-500/20'
-            )}>
-              <div className="flex items-start gap-3">
-                <GitBranch className={cn('w-8 h-8 mt-1', driftedResources > 0 ? 'text-red-400' : 'text-green-400')} />
-                <div className="flex-1 min-w-0">
-                  <h3 className="text-lg font-semibold text-foreground">
-                    {driftedResources > 0 ? 'Configuration Drift Detected' : 'No Drift Detected'}
-                  </h3>
-                  <p className="text-sm text-muted-foreground mt-1">
-                    {driftedResources > 0
-                      ? `${driftedResources} resource(s) have drifted from the desired Git state`
-                      : 'Cluster configuration matches Git repository state'}
-                  </p>
-                  <div className="flex flex-wrap gap-4 mt-3 text-sm text-muted-foreground">
-                    {gitRepo && (
-                      <div className="flex items-center gap-1.5">
-                        <GitBranch className="w-4 h-4" />
-                        <span>{gitRepo}</span>
-                      </div>
-                    )}
-                    {gitBranch && (
-                      <div className="flex items-center gap-1.5">
-                        <ArrowRight className="w-4 h-4" />
-                        <span>Branch: {gitBranch}</span>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            {/* Quick Stats */}
-            <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-              <div className="p-4 rounded-lg border border-border bg-card/50">
-                <div className={cn('text-2xl font-bold', severityStyle.text)}>
-                  <SeverityIcon className="w-8 h-8" />
-                </div>
-                <div className="text-xs text-muted-foreground mt-1">{t('common.status')}</div>
-              </div>
-              <div className="p-4 rounded-lg border border-border bg-card/50">
-                <div className={cn('text-2xl font-bold', driftedResources > 0 ? 'text-red-400' : 'text-green-400')}>
-                  {driftedResources}
-                </div>
-                <div className="text-xs text-muted-foreground">{t('drilldown.drift.driftedResources')}</div>
-              </div>
-              <div className="p-4 rounded-lg border border-border bg-card/50">
-                <div className="text-sm font-mono text-foreground truncate">{gitPath || '/'}</div>
-                <div className="text-xs text-muted-foreground">{t('drilldown.drift.gitPath')}</div>
-              </div>
-              <div className="p-4 rounded-lg border border-border bg-card/50">
-                <div className="text-sm text-foreground">
-                  {lastChecked ? new Date(lastChecked).toLocaleString() : '-'}
-                </div>
-                <div className="text-xs text-muted-foreground">{t('drilldown.drift.lastChecked')}</div>
-              </div>
-            </div>
-
-            {/* Git Source Info */}
-            {gitRepo && (
-              <div className="p-4 rounded-lg border border-border bg-card/50">
-                <h4 className="text-sm font-medium text-foreground mb-3">{t('drilldown.drift.gitSource')}</h4>
-                <div className="grid grid-cols-2 gap-4 text-sm">
-                  <div>
-                    <span className="text-muted-foreground">Repository:</span>
-                    <span className="ml-2 text-foreground font-mono">{gitRepo}</span>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground">Branch:</span>
-                    <span className="ml-2 text-foreground">{gitBranch || 'main'}</span>
-                  </div>
-                  <div>
-                    <span className="text-muted-foreground">Path:</span>
-                    <span className="ml-2 text-foreground font-mono">{gitPath || '/'}</span>
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
+          <DriftOverviewTab
+            driftedResources={driftedResources}
+            driftSeverity={driftSeverity}
+            gitRepo={gitRepo}
+            gitBranch={gitBranch}
+            gitPath={gitPath}
+            lastChecked={lastChecked}
+            severityStyle={severityStyle}
+          />
         )}
-
         {activeTab === 'changes' && (
-          <div className="space-y-4">
-            <h4 className="text-sm font-medium text-foreground">Drifted Resources ({changes?.length || 0})</h4>
-            {changesError && (
-              <div className="flex items-center gap-2 p-3 rounded-lg bg-red-500/10 border border-red-500/30 text-red-300">
-                <AlertTriangle className="w-4 h-4 shrink-0" />
-                <span className="text-sm">{changesError}</span>
-              </div>
-            )}
-            {changesLoading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-muted-foreground" />
-              </div>
-            ) : changes && changes.length > 0 ? (
-              <div className="space-y-2">
-                {changes.map((change, i) => {
-                  const changeStyle = getChangeTypeStyle(change.changeType)
-                  return (
-                    <div
-                      key={i}
-                      onClick={() => {
-                        handleResourceClick(change)
-                        setSelectedChange(change)
-                      }}
-                      className={cn(
-                        'flex items-center justify-between p-3 rounded-lg border border-border bg-card/50',
-                        (change.kind === 'Pod' || change.kind === 'Deployment') && change.namespace
-                          ? 'cursor-pointer hover:bg-card/80 transition-colors'
-                          : ''
-                      )}
-                    >
-                      <div className="flex items-center gap-3">
-                        <span className={cn('px-2 py-0.5 rounded text-xs font-medium', changeStyle.bg, changeStyle.text)}>
-                          {changeStyle.label}
-                        </span>
-                        <div>
-                          <span className="text-sm font-medium text-foreground">{change.kind}/{change.name}</span>
-                          {change.namespace && (
-                            <span className="text-xs text-muted-foreground ml-2">({change.namespace})</span>
-                          )}
-                        </div>
-                      </div>
-                      {(change.kind === 'Pod' || change.kind === 'Deployment') && change.namespace && (
-                        <svg className="w-4 h-4 text-muted-foreground" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-                        </svg>
-                      )}
-                    </div>
-                  )
-                })}
-              </div>
-            ) : (
-              <div className="text-center py-12 text-muted-foreground">
-                <CheckCircle className="w-12 h-12 mx-auto mb-3 opacity-50 text-green-400" />
-                <p className="text-green-400">{t('drilldown.drift.noDrifted')}</p>
-                <p className="text-xs mt-1">{t('drilldown.drift.allMatch')}</p>
-              </div>
-            )}
-          </div>
+          <DriftChangesTab
+            changes={changes}
+            changesLoading={changesLoading}
+            changesError={changesError}
+            selectedChange={selectedChange}
+            onChangeClick={handleResourceClick}
+            getChangeTypeStyle={getChangeTypeStyle}
+          />
         )}
-
         {activeTab === 'diff' && (
-          <div className="space-y-4">
-            <h4 className="text-sm font-medium text-foreground">{t('drilldown.drift.configDiff')}</h4>
-            {selectedChange ? (
-              <div className="space-y-4">
-                <div className="p-3 rounded-lg bg-secondary/50">
-                  <span className="text-sm text-foreground">
-                    {selectedChange.kind}/{selectedChange.name}
-                    {selectedChange.namespace && ` in ${selectedChange.namespace}`}
-                  </span>
-                </div>
-                {selectedChange.diff ? (
-                  <div className="p-4 rounded-lg border border-border bg-card/50">
-                    <pre className="text-sm font-mono whitespace-pre-wrap overflow-x-auto">
-                      {selectedChange.diff}
-                    </pre>
-                  </div>
-                ) : selectedChange.fields && selectedChange.fields.length > 0 ? (
-                  <div className="space-y-2">
-                    {selectedChange.fields.map((field, i) => (
-                      <div key={i} className="p-3 rounded-lg border border-border bg-card/50">
-                        <div className="text-xs text-muted-foreground mb-2 font-mono">{field.path}</div>
-                        <div className="grid grid-cols-2 gap-4">
-                          <div>
-                            <span className="text-xs text-muted-foreground">Git:</span>
-                            <pre className="text-sm text-green-400 mt-1">{field.gitValue}</pre>
-                          </div>
-                          <div>
-                            <span className="text-xs text-muted-foreground">{t('drilldown.fields.cluster')}</span>
-                            <pre className="text-sm text-red-400 mt-1">{field.clusterValue}</pre>
-                          </div>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="text-center py-12 text-muted-foreground">
-                    <FileText className="w-12 h-12 mx-auto mb-3 opacity-50" />
-                    <p>{t('drilldown.drift.diffNotAvailable')}</p>
-                    <p className="text-xs mt-1">{t('drilldown.drift.selectResourceChanges')}</p>
-                  </div>
-                )}
-              </div>
-            ) : (
-              <div className="text-center py-12 text-muted-foreground">
-                <Diff className="w-12 h-12 mx-auto mb-3 opacity-50" />
-                <p>{t('drilldown.drift.selectResourceDiff')}</p>
-                <p className="text-xs mt-1">{t('drilldown.drift.chooseDrifted')}</p>
-              </div>
-            )}
-          </div>
+          <DriftDiffTab selectedChange={selectedChange} />
         )}
-
         {activeTab === 'ai' && (
-          <div className="space-y-4">
-            <div className="flex items-center justify-between">
-              <h4 className="text-sm font-medium text-foreground flex items-center gap-2">
-                <ConsoleAIIcon className="w-5 h-5" />
-                {t('drilldown.ai.title')}
-              </h4>
-              <button
-                onClick={handleDiagnose}
-                disabled={!isAgentConnected}
-                className="flex items-center gap-2 px-3 py-1.5 rounded-lg bg-purple-500/20 text-purple-400 hover:bg-purple-500/30 disabled:opacity-50 disabled:cursor-not-allowed transition-colors text-sm"
-              >
-                <Stethoscope className="w-4 h-4" />
-                Analyze Drift
-              </button>
-            </div>
-
-            {!isAgentConnected ? (
-              <div className="text-center py-12 text-muted-foreground">
-                <ConsoleAIIcon className="w-12 h-12 mx-auto mb-3 opacity-50" />
-                <p>{t('drilldown.ai.notConnected')}</p>
-                <p className="text-xs mt-1">{t('drilldown.ai.configureAgent')}</p>
-              </div>
-            ) : aiAnalysisLoading ? (
-              <div className="flex items-center justify-center py-12">
-                <Loader2 className="w-6 h-6 animate-spin text-purple-400" />
-              </div>
-            ) : aiAnalysis ? (
-              <div className="p-4 rounded-lg bg-purple-500/10 border border-purple-500/20">
-                <pre className="whitespace-pre-wrap text-sm text-foreground">{aiAnalysis}</pre>
-              </div>
-            ) : (
-              <div className="text-center py-12 text-muted-foreground">
-                <Stethoscope className="w-12 h-12 mx-auto mb-3 opacity-50" />
-                <p>{t('drilldown.drift.clickAnalyze')}</p>
-                <p className="text-xs mt-1">{t('drilldown.drift.analyzeHint')}</p>
-              </div>
-            )}
-          </div>
+          <DriftAITab
+            isAgentConnected={isAgentConnected}
+            aiAnalysis={null}
+            aiAnalysisLoading={false}
+            onDiagnose={handleDiagnose}
+          />
         )}
       </div>
     </div>
